fix issues 3 (move questions and ansers features in different specs)
