+++ spec/acceptance/any_user_can_see_the_list_of_questions_spec.rb
@@ -1,8 +1,8 @@
 require 'rails_helper'
 
-feature 'See all questions', %q{
-  any user can see the questions index view page
-} do
+feature 'SEE ALL QUESTIONS', %q(
+  any user can see the questions index page
+) do
   given(:user) { create(:user) }
   given!(:questions) { create_list(:question, 5, user: user) }
